+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/search/SearchDialogFragment.kt
@@ -66,8 +66,10 @@ import mozilla.components.support.ktx.android.content.hasCamera
 import mozilla.components.support.ktx.android.content.isPermissionGranted
 import mozilla.components.support.ktx.android.content.res.getSpanned
 import mozilla.components.support.ktx.android.net.isHttpOrHttps
+import mozilla.components.support.ktx.android.view.ImeInsetsSynchronizer
 import mozilla.components.support.ktx.android.view.findViewInHierarchy
 import mozilla.components.support.ktx.android.view.hideKeyboard
+import mozilla.components.support.ktx.android.view.setupPersistentInsets
 import mozilla.components.support.ktx.android.view.showKeyboard
 import mozilla.components.support.ktx.kotlin.toNormalizedUrl
 import mozilla.components.ui.autocomplete.InlineAutocompleteEditText
@@ -193,6 +195,7 @@ class SearchDialogFragment : AppCompatDialogFragment(), UserInteractionHandler {
             if ((requireActivity() as HomeActivity).browsingModeManager.mode.isPrivate) {
                 this.secure(requireActivity())
             }
+            window?.setupPersistentInsets()
         }
     }
 
@@ -456,6 +459,7 @@ class SearchDialogFragment : AppCompatDialogFragment(), UserInteractionHandler {
             updateAccessibilityTraversalOrder()
         }
 
+        ImeInsetsSynchronizer.setup(view)
         observeClipboardState()
         observeAwesomeBarState()
         observeSuggestionProvidersState()